requirements.txt
django==5.1.2
python-dotenv>=1.0.1,<2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
qrcode==7.4.2
Pillow==10.4.0

django-filter==24.3
drf-spectacular==0.27.2
firebase-admin==6.5.0
cryptography==43.0.0
stripe==8.9.0

# Production dependencies
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9  # PostgreSQL adapter for Django

# Media storage (for production - prevents file loss on Scalingo)
django-storages==1.14.2
cloudinary==1.41.0
django-cloudinary-storage==0.3.0

# External APIs
requests==2.31.0
twilio==9.3.0

# Backups (S3 upload support)
boto3==1.34.162

# PDF Generation
reportlab==4.0.9
